requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.0
cryptography==42.0.7
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fiona==1.9.6
Flask==2.3.3
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
GeoAlchemy2==0.15.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.2.1
iniconfig==2.0.0
install==1.3.5
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpath-ng==1.6.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
llama-index==0.10.36
llama-index-agent-openai==0.2.1
llama-index-cli==0.1.11
llama-index-core==0.10.37
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.19
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.12
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.0
llamaindex-py-client==0.1.19
loguru==0.7.2
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
more-itertools==10.2.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.30.1
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.1
pluggy==1.5.0
ply==3.11
preshed==3.0.9
probableparsing==0.0.1
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.2
pypdf==4.1.0
pyproj==3.6.1
pytest==8.1.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ratelim==0.1.6
regex==2023.12.25
requests==2.31.0
ruff==0.4.3
sentry-sdk==2.1.1
shapely==2.0.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlacodegen==2.3.0
SQLAlchemy==2.0.25
sqlalchemy-codegen==1.1.2
srsly==2.4.8
stack-data==0.6.3
striprtf==0.0.26
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tqdm==4.66.2
traitlets==5.14.2
typeguard==4.2.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
usaddress==0.5.10
usaddress-scourgify==0.6.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0
yaml-config==0.1.5
yarl==1.9.4
zipp==3.18.1
paramiko~=3.4.0

